safari-534.53-branch/LayoutTests/http/tests/security/MessagePort/event-listener-context-expected.txt

Test that JS code cannot be planted into an unsuspecting subframe via MessagePort and navigation.

PASS: got exception while creating port from closed frame's constructor: ReferenceError: MessageChannel constructor associated document is unavailable
DONE
